+++ src/neomesh.h
@@ -184,7 +184,7 @@ private:
     void login_sapi(uint8_t * password);
     void change_node_id_sapi(uint16_t nodeid);
     void write_raw(uint8_t *data, uint8_t length);
-    bool wait_for_sapi_response(tNcSapiMessage * message, uint16_t timeout_ms);
+    bool wait_for_sapi_response(tNcSapiMessage * message, uint32_t timeout_ms);
     void start_bootloader();
     void start_protocol_stack();
     uint8_t get_setting(uint8_t setting);
